Moved handling of adding/removing tree nodes to onFlush() hook
All job previously done in prePersist() hook is now called in onFlush()
and is called "processScheduledInsertion()". Same for removal.

prePersist and preRemove hooks are still called for compatibility, but
as for now they do nothing both for Nested and Closure Strategy

Added unit test to ensure early persist is working correctly