Removed the Urlizer class and reuse behat/transliterator
Both classes are exactly the same. Behat and the Gedmo extension have both
copied the Doctrine 1 code in their own project. For Behat 3, the code has
been moved to a package on its own to allow reusability with composer. So
we can now depend on it instead of maintaining a separate copy of the same
code.