Fix translation handling when modifying documents
When modifying a new document within a flush listener that occurs before the
TranslatableListener, the TranslatableListener will see an insert as well as an
update. This causes duplicate key exceptions since a previous persist() is only
checked for personal translation documents, not for the default class. This
commit changes the behavior so that previously inserted persisted translation
documents that have not yet been written to the database are simply updated.