Fix the id strategy for closure and translation entities
The Closure strategy and the translatable listener are inserting the
data in SQL without providing the identifier, assuming it is generated
by the database. However, the default id strategy for Doctrine is
'AUTO', which chooses the strategy depending on the platform. When the
sequence strategy is preferred (PostgreSQL and Oracle), these
listeners were broken. Forcing the strategy to IDENTITY rather than AUTO
fixes it.