Persistance of position relocations moved to other events to fix issue #1209
Added an unit test for issue #1209 and moved the persistance of position relocations from onFlush event, which is called before database transaction started, to the first call of postPersist, preUpdate or preRemove. So that the update query is executed within the transaction and if errors happen during queries the position changes will be rolled back. The synchronization of the objects has been separated from the database persistance, so that it will happen only if all queries were successfully executed.