Use mb_strlen to get string length
I use the sluggable extension to get a user-friendly path generated for my tree nodes.
For that, the name used to generate the slug is the unaltered name of the node (which can contain accentuated characters). These characters are counted as more than 1 character by `strlen`, which messes up the result of `replaceRelative`. Using `mb_strlen` instead solves the problem!