[Translatable] Fix replacement of last group by element
Some Sql vendors (like Sql Server) require every selected element to be in
the group by clause. Because the translation walker adds additional
columns, those columns must be added to the group by as well. The regex was incorrect and did not replace the very
last element in the group by clause.