[Sortable] Fix issue with position synchronization after multiple deletes
In the `postFlush` method of the `SortableListener` the position of objects get updated via `$this->setFieldValue()` after applying the relocation. The problem is that `$this->setFieldValue()` creates a change set for the passed object and that will prevent from other relocations being executed on this object.

I have added a unit test that tests this case and extended the behaviour of `postFlush` so that all relocations for one object will get applied.