Fix Travis overwritten env
Env variable works well, was just getting overwritten for the lowest dependency tests. Add the new var to all envs and we should be good.

Thought about removing the tests for lowest dependency, or make them ALL low dependency, but I think this adds good potential coverage, especially for a package of this size.