No longer test lowest dependencies
Travis builds have been failing since Travis updated to MongoDB 4.0. Somewhere in an old release of a “require-dev” dependency, likely Doctrine’s ODM (or one of its dependencies), a deprecated MongoDB feature called “$pushAll” is still being used.

This deprecated feature isn’t used on the latest version of dev dependencies, so somewhere in there it was bug fixed or whatever.

Bumping the minimum version requirements in composer.json would be a breaking change and require a new major release. While some of this is already being worked on for Extensions 3.0, we would like to have normal CI tests for 2.4.x in the meantime. That’s what this hopes to accomplish.