Update MongoDB ODM to 2.0, require Doctrine Mongo ext
The Doctrine MongoDB package used to be required by ODM 1.3. ODM 2.0 makes use of the official MongoDB adapter package, and no longer requires their own adapter layer.

Doctrine Extensions does use the adapter layer features in places, so we still need it. We will want to remove the dependency in a future release, as Doctrine MongoDB package is in maintainace mode only, and not actively supported.