Resolve Column Type when fetching Max Position from Database
In SQLite it makes a difference if date columns are queried with '00:00:00' as timestring or without. MySQL seems to ignore the time completely. Unfortunately there is an asymmetry between SQL Types and PHP Objects: 'date' and 'datetime' Column point both to a DateTime Object in PHP. Therefore, we cannot derive the column type from the PHP Object. Luckily we can ask the ClassMetadata in Doctrine for the right type.