Documented the usage of refresh query hint if cache is used for translatable (#1021)
If query caching is enabled the TranslationWalker won't be used anymore once the query is cached. Therefore the `Query::HINT_REFRESH` won't be set automatically. This leads to the issue, that once a translateable entity is loaded a second query with another locale won't update the translateable properties in the entity.

Adding the `Query::HINT_REFRESH` manually to the query will fix the issue.